Remove agentops.enums module; move EndState to .session, EventType to .event
